+++ rust/operator-binary/src/trino_controller.rs
@@ -331,6 +331,7 @@ mod tests {
             trino_opa_config,
             resolved_fte_config,
             resolved_client_protocol_config,
+            resolved_lineage_config: None,
         };
 
         let operator_env = OperatorEnvironmentOptions {
@@ -562,6 +563,7 @@ mod tests {
             trino_opa_config: None,
             resolved_fte_config: None,
             resolved_client_protocol_config: None,
+            resolved_lineage_config: None,
         };
         let operator_env = OperatorEnvironmentOptions {
             operator_namespace: "stackable-operators".to_string(),